+++ src/main/java/world/trecord/service/record/RecordService.java
@@ -55,7 +55,7 @@ public class RecordService {
     public RecordCreateResponse createRecord(Long userId, RecordCreateRequest request) {
         FeedEntity feedEntity = findFeedOrException(request.getFeedId());
 
-        doEnsureUserHasPermissionOverFeed(feedEntity, userId);
+        ensureUserHasPermissionOverFeed(feedEntity, userId);
 
         int nextSequence = findNextSequence(feedEntity.getId(), request.getDate());
         RecordEntity savedRecordEntity = recordRepository.save(request.toEntity(feedEntity, nextSequence));
@@ -71,7 +71,7 @@ public class RecordService {
         RecordEntity recordEntity = findRecordForUpdateOrException(recordId);
         FeedEntity feedEntity = findFeedOrException(recordEntity.getFeedEntity().getId());
 
-        doEnsureUserHasPermissionOverFeed(feedEntity, userId);
+        ensureUserHasPermissionOverFeed(feedEntity, userId);
 
         recordEntity.update(request.toUpdateEntity());
         recordRepository.saveAndFlush(recordEntity);
@@ -92,7 +92,7 @@ public class RecordService {
 
         FeedEntity feedEntity = findFeedOrException(originalRecord.getFeedEntity().getId());
 
-        doEnsureUserHasPermissionOverFeed(feedEntity, userId);
+        ensureUserHasPermissionOverFeed(feedEntity, userId);
 
         originalRecord.swapSequenceWith(targetRecord);
         recordRepository.saveAllAndFlush(List.of(originalRecord, targetRecord));
@@ -107,7 +107,7 @@ public class RecordService {
     public void deleteRecord(Long userId, Long recordId) {
         RecordEntity recordEntity = findRecordForUpdateOrException(recordId);
         FeedEntity feedEntity = findFeedOrException(recordEntity.getFeedEntity().getId());
-        doEnsureUserHasPermissionOverFeed(feedEntity, userId);
+        ensureUserHasPermissionOverFeed(feedEntity, userId);
 
         commentRepository.deleteAllByRecordEntityId(recordId);
         userRecordLikeRepository.deleteAllByRecordEntityId(recordId);
@@ -152,7 +152,7 @@ public class RecordService {
         return recordRepository.findByIdForUpdate(recordId).orElseThrow(() -> new CustomException(RECORD_NOT_FOUND));
     }
 
-    private void doEnsureUserHasPermissionOverFeed(FeedEntity feedEntity, Long userId) {
+    private void ensureUserHasPermissionOverFeed(FeedEntity feedEntity, Long userId) {
         if (!feedEntity.isManagedBy(userId)) {
             throw new CustomException(FORBIDDEN);
         }
